New Commit 30May2023

+++ EmployeeApplication_WinForm/EmployeeApplication_WinForm/EmployeeData.cs
@@ -15,6 +15,7 @@ namespace EmployeeApplication_WinForm
 {
     public partial class EmployeeData : Form
     {
+        EmployeeModel employeeModel = new EmployeeModel();
         public EmployeeData()
         {
             InitializeComponent();
@@ -25,41 +26,94 @@ namespace EmployeeApplication_WinForm
         {
 
             var response = await RestApicall.GetAll();
-         
-            txtView.Text =RestApicall.EmpJson(response);
+            if (txtEmpName.Text==""&txtId.Text==""&txtEmail.Text==""&txtGender.Text==""&txtStatus.Text=="")
+            {
+                txtView.Text = RestApicall.EmpJson(response);
+
+            }
+            else
+            {
+                MessageBox.Show("Please Don't Insert The Employee Details Feilds");
+            }
             ClearData();
         }
 
        
         private async void btnView_Click(object sender, EventArgs e)
         {
-            var response = await RestApicall.GetById(txtId.Text);
-            txtView.Text = RestApicall.EmpJson(response);
+            if(txtId.Text!=""& txtEmail.Text == "" & txtGender.Text == "" & txtStatus.Text == ""&txtEmpName.Text=="")
+            {
+                var response = await RestApicall.GetById(txtId.Text);
+                txtView.Text = RestApicall.EmpJson(response);
+
+            }
+            else
+            {
+                txtView.Text = "";
+                MessageBox.Show("Please Insert Employee Id Feild Only");
+            }
             ClearData();
         }
 
         private async void btnAddEmp_Click(object sender, EventArgs e)
         {
-            var response = await RestApicall.AddEmployee(txtEmpName.Text,txtEmail.Text,txtGender.Text,txtStatus.Text);
-            txtView.Text = RestApicall.EmpJson(response);
-            MessageBox.Show("Employee Data Added Sucessfully");
+            if ( txtEmail.Text != "" & txtGender.Text != "" & txtStatus.Text != "" & txtEmpName.Text != "")
+            {
+                if(txtId.Text == "")
+                {
+                    var response = await RestApicall.AddEmployee(txtEmpName.Text, txtEmail.Text, txtGender.Text, txtStatus.Text);
+                    txtView.Text = RestApicall.EmpJson(response);
+                    MessageBox.Show("Employee Data Added Sucessfully");
+                }
+                else
+                {
+                    MessageBox.Show("Please Don't enter Employee Id Feild");
+                }
+            }
+            else
+            {
+                txtView.Text = "";
+                MessageBox.Show("Please Enter EmployeeName,Email,Gender,Status Feilds");
+
+            }
             ClearData();
         }
 
         private async void btnUpdateEmployee_Click(object sender, EventArgs e)
         {
-            var response = await RestApicall.UpdateEmployee(txtId.Text,txtEmpName.Text, txtEmail.Text, txtGender.Text, txtStatus.Text);
-            txtView.Text = RestApicall.EmpJson(response);
-            MessageBox.Show("Employee Data Updated Sucessfully");
+
+            if (txtId.Text != "" &txtEmail.Text != "" & txtGender.Text != "" & txtStatus.Text != "" & txtEmpName.Text != "")
+            {
+               
+
+                    var response = await RestApicall.UpdateEmployee(txtId.Text, txtEmpName.Text, txtEmail.Text, txtGender.Text, txtStatus.Text);
+                    txtView.Text = RestApicall.EmpJson(response);
+                    
+                    MessageBox.Show("Employee Data Updated Sucessfully");
+                
+            }
+
+            else
+            {
+                MessageBox.Show("Please Enter EmployeeId, EmployeeName,Email,Gender,Status Feilds");
+            }
 
             ClearData();
         }
 
         private async void btnDeleteEmp_Click(object sender, EventArgs e)
         {
+            if (txtId.Text != "" & txtEmail.Text == "" & txtGender.Text == "" & txtStatus.Text == "" & txtEmpName.Text == "")
+            {
 
-            var response = await RestApicall.DeleteEmployee(txtId.Text);
-            txtView.Text = response;
+                var response = await RestApicall.DeleteEmployee(txtId.Text);
+                txtView.Text = response;
+            }
+            else
+            {
+                MessageBox.Show("Please Enter only EmployeeId");
+                txtView.Text = "";
+            }
             ClearData();
         }
